neovim-options.lua: comment out splitbelow and splitright
Use the default behaviour.

Similar to dwm (Dynamic Window Manager) it makes sense that a new split
is placed at the top/left because it brings the new window to your focus
(usually top/left) and because what you plan to do in that new window
(search, edit, etc.) that is more important than what you're doing in
the previous window. Usually, I split windows to look for a definition,
search, open a new file, etc., to compare/confirm something related to
the current window. That means that MOST of the time I split windows
temporarily. This adapts to my current
workflow.